feat: add agent engine deployment support with environment configuration
