+++ .env.example
@@ -1,0 +1,6 @@
+GCLOUD_LOCATION=us-central1
+GCLOUD_PROJECT_ID={your-project-id}
+GOOGLE_GENAI_USE_VERTEXAI=TRUE
+STAGING_BUCKET=gs://purchasing-concierge-{your-project-id}
+PIZZA_SELLER_AGENT_URL={your-pizza-agent-url}
+BURGER_SELLER_AGENT_URL={your-burger-agent-url}
