+++ pyproject.toml
@@ -6,6 +6,7 @@ readme = "README.md"
 requires-python = ">=3.12"
 dependencies = [
     "google-adk[a2a]>=1.8.0",
+    "google-cloud-aiplatform[agent-engines]>=1.105.0",
     "gradio>=5.28.0",
     "httpx>=0.28.1",
     "jwcrypto>=1.5.6",
